cli: move task logging utils to "new" module
