+++ packages/cli/src/modules/new/lib/execution/installNewPackage.ts
@@ -17,7 +17,7 @@ import fs from 'fs-extra';
 import upperFirst from 'lodash/upperFirst';
 import camelCase from 'lodash/camelCase';
 import { paths } from '../../../../lib/paths';
-import { Task } from '../../../../lib/tasks';
+import { Task } from '../tasks';
 import { PortableTemplateInput } from '../types';
 
 export async function installNewPackage(input: PortableTemplateInput) {
